Add warning to non pickable models
